Create add-post form handler
