+++ add.php
@@ -3,7 +3,7 @@
 require_once 'utils/constants.php';
 require_once 'utils/helpers.php';
 require_once 'utils/functions.php';
-require_once 'utils/post-form-validators.php';
+require_once 'utils/form-handlers/add-post.php';
 require_once 'utils/renderers/common.php';
 require_once 'models/content_type.php';
 require_once 'models/post.php';
@@ -15,9 +15,7 @@ require_once 'init/db-connection.php';
  * @var mysqli $db_connection - ресурс соединения с базой данных
  */
 
-$form_data = [
-    'author_id' => $user_session['id'],
-];
+$form_data = [];
 $errors = [];
 
 $layout_data = [
@@ -27,80 +25,17 @@ $layout_data = [
 ];
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    $with_file = isset($_FILES['photo-file'])
-                 && $_FILES['photo-file']['error'] !== UPLOAD_ERR_NO_FILE;
-
-    $form_data['content_type_id'] = $_POST['content-type-id'] ?? '';
-    $form_data['title'] = $_POST['title'] ?? '';
-    $form_data['text_content'] = $_POST['text-content'] ?? '';
-    $form_data['string_content'] =
-        !$with_file ? $_POST['string-content'] ?? '' : '';
-    $form_data['tags'] =
-        $_POST['tags'] ? trim(
-            preg_replace('/\s+/', TEXT_SEPARATOR, mb_strtolower($_POST['tags']))
-        ) : '';
-    $form_data['photo_file'] =
-        $with_file ? $_FILES['photo-file'] : null;
-
-    $content_type_data = $form_data['content_type_id'] ? get_content_type(
-        $db_connection,
-        $form_data['content_type_id']
-    ) : null;
-    $content_type = $content_type_data && $content_type_data['type']
-        ? $content_type_data['type'] : null;
-
-    $errors = $content_type
-        ? get_post_form_data_errors($form_data, $content_type)
-        : [
-            [
-                'title' => 'Тип контента',
-                'description' => 'Некорректный тип'
-            ]
-        ];
-
-    if (count($errors)) {
-        if ($with_file && !$errors['photo_file']) {
-            $errors['photo_file'] = [
-                'title' => 'Файл фото',
-                'description' => 'Загрузите файл еще раз'
-            ];
-        }
-    }
-
-    $is_photo_content_type = $content_type === 'photo';
-    $photo_url = '';
-
-    if (!count($errors) && $is_photo_content_type) {
-        $photo_url =
-            $with_file
-                ? save_file($form_data['photo_file'])
-                : download_file($form_data['string_content']);
-
-        if (!$photo_url) {
-            if ($with_file) {
-                $errors['photo_file'] = [
-                    'title' => 'Файл фото',
-                    'description' => 'Не удалось загрузить файл'
-                ];
-            } else {
-                $errors['string_content'] = [
-                    'title' => 'Ссылка из интернета',
-                    'description' => 'Не удалось загрузить файл по ссылке'
-                ];
-            }
-        }
-    }
+    list(
+        'form_data' => $form_data,
+        'errors' => $errors
+        ) = handle_add_post_form($db_connection);
 
     if (!count($errors)) {
-        if ($is_photo_content_type) {
-            $form_data['string_content'] = $photo_url;
-        }
-
+        $form_data['author_id'] = $user_session['id'];
         $created_post_id = create_post($db_connection, $form_data);
 
         if ($created_post_id) {
             header("Location: post.php?post_id=$created_post_id");
-
             exit();
         }
