+++ utils/form-handlers/login.php
@@ -2,6 +2,7 @@
 
 require_once 'utils/constants.php';
 
+// todo: update phpDoc
 /**
  * Функция обрабабатыват данные формы аутентификации.
  * В случае успешной аутентификации происходит
